pyproject.toml
[project]
name = "arctic-defense-scraper"
version = "1.0.0"
description = "Automated scraper for arctic-relevant defense opportunities"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.5.0",
    "aiohttp>=3.9.0",
    "playwright>=1.40.0",
    "google-genai>=1.0.0",
    "python-dateutil>=2.8.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "tenacity>=8.2.0",
    "aiofiles>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
